generative-ai/llm/2-doc-examples/2-2-langchain/4-tutorials/build-simple-llm-app/pyproject.toml
[project]
name = "build-simple-llm-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.2.8",
    "langchain>=0.3.7",
    "python-dotenv>=1.0.1",
]
